// src/app/api/fn/reports/reports-receivables-retrieve.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { Receivables } from '../../models/receivables';

export interface ReportsReceivablesRetrieve$Params {

/**
 * overdue, over_limit or on_hold; everyone who owes when left out.
 */
  show?: 'on_hold' | 'over_limit' | 'overdue';
}

export function reportsReceivablesRetrieve(http: HttpClient, rootUrl: string, params?: ReportsReceivablesRetrieve$Params, context?: HttpContext): Observable<StrictHttpResponse<Receivables>> {
  const rb = new RequestBuilder(rootUrl, reportsReceivablesRetrieve.PATH, 'get');
  if (params) {
    rb.query('show', params.show, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<Receivables>;
    })
  );
}

reportsReceivablesRetrieve.PATH = '/api/reports/receivables/';